Split plot functions into multiple files
