+++ docs/python_api/index.md
@@ -109,6 +109,12 @@ The **dynamiqs** Python API features two main types of functions: solvers of dif
 
 ### Plotting
 
-::: dynamiqs.plots.misc
+::: dynamiqs.plots.namespace
     options:
         table: true
+        members:
+        - plot_wigner
+        - plot_wigner_mosaic
+        - plot_pwc_pulse
+        - plot_fock
+        - plot_fock_evolution
