+++ dynamiqs/plots/namespace.py
@@ -1,0 +1,11 @@
+from .plots_fock import plot_fock, plot_fock_evolution
+from .plots_misc import plot_pwc_pulse
+from .plots_wigner import plot_wigner, plot_wigner_mosaic
+
+__all__ = [
+    'plot_wigner',
+    'plot_wigner_mosaic',
+    'plot_pwc_pulse',
+    'plot_fock',
+    'plot_fock_evolution',
+]
