+++ mkdocs.yml
@@ -194,8 +194,8 @@ nav:
               - snap_gate: python_api/utils/optimal_control/snap_gate.md
               - cd_gate: python_api/utils/optimal_control/cd_gate.md
           - Plotting:
-              - plot_wigner: python_api/plots/misc/plot_wigner.md
-              - plot_wigner_mosaic: python_api/plots/misc/plot_wigner_mosaic.md
-              - plot_pwc_pulse: python_api/plots/misc/plot_pwc_pulse.md
-              - plot_fock: python_api/plots/misc/plot_fock.md
-              - plot_fock_evolution: python_api/plots/misc/plot_fock_evolution.md
+              - plot_wigner: python_api/plots/namespace/plot_wigner.md
+              - plot_wigner_mosaic: python_api/plots/namespace/plot_wigner_mosaic.md
+              - plot_pwc_pulse: python_api/plots/namespace/plot_pwc_pulse.md
+              - plot_fock: python_api/plots/namespace/plot_fock.md
+              - plot_fock_evolution: python_api/plots/namespace/plot_fock_evolution.md
